Gardening: Remove unnecessary FIXME from GeoNotifier
GeoNotifier is managed in Geolocation and it's only used in Geolocation
and GeolocationWatcher which is also managed in Geolocation.
Therefore m_geolocation in GeoNotifier is not a problem
because there is no possibility to be used without Geolocation.

BUG=

Review URL: https://codereview.chromium.org/1381563006

Cr-Original-Commit-Position: refs/heads/master@{#353482}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 355886195c74db484629c0d1a7113ab9751b39af