Clear distributions when inserting an InsertionPoint to avoid cycles.
We need to clear the distributions from insertion points both in removedFrom
to avoid keeping otherwise unreferenced elements from being GC'ed, and also
in insertedInto so that distributions from detached subtrees don't result in
cycles during the detach step of lazyReattachIfAttached during the later
distribution.

Previously we were not clearing it in insertedInto, which meant if you had an
element A with a child B, and distributed B into a <content> inside A's
ShadowRoot, then removed B, put it into the page, and inserted A as a child of
it, you could produce a situation where B was still a child of A in the
distribution lists, but inside the DOM B is actually the parent of A. This then
results in infinite recursion where InsertionPoint::detach would actually go
to an ancestor of the insertion point itself.

BUG=541533

Review URL: https://codereview.chromium.org/1395893005

Cr-Original-Commit-Position: refs/heads/master@{#353704}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 56f466b6f6b8e5225af59f61dba452a98a400d06