Add a willUseRange notifier for FontSelector
For improving font fallback in the complex code path we need better
control over which unicode-range out of a set of CSSFontFaces are
getting loaded. Just a UChar32 is not sufficient for that purpose. So
identify the face to be loaded by the range parameters for now.

BUG=542629
R=eae,leviw,kojii

Review URL: https://codereview.chromium.org/1401883004

Cr-Original-Commit-Position: refs/heads/master@{#353730}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 88480e6547903613cfce0473916c6768a25bdded