Inherit vertical-align to text nodes unless it is anonymous inline
This patch supports anonymous inline boxes[1] in terms of inheritance
for the vertical-align property.

crbug.com/496850 changed the vertical-align property to compute to the
initial value for any text nodes, while the spec defines to do so only
when it is an anonymous inline. This patch works for most cases,
because there is an optimized code path when
descendantsHaveSameLineHeightAndBaseline() is true[2], and it is true
in most cases where a text node is not an anonymous inline.

There are some cases that it is false, and the fix caused the
vertical-align not applied properly in those cases, such as fallback
fonts are used. This patch fixes such cases.

[1] http://www.w3.org/TR/CSS21/visuren.html#anonymous
[2] https://trac.webkit.org/changeset/82280/trunk/Source/WebCore/rendering/RenderBlockLineLayout.cpp

BUG=537277, 496850

Review URL: https://codereview.chromium.org/1372413003

Cr-Original-Commit-Position: refs/heads/master@{#353984}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 055db165e126fb377b7483758616212911e992da