Add a test to confirm that multiple missing images in webkit-mask-image prevent rendering
When specifying more than one missing image in webkit-mask-image rule,
the element for which the style is applied is not rendered correctly.
If only one missing image is specified, the element is rendered.

This patch migrates the relevant Manual Test to the Layout Test folder
and adds a failing test expectation for the test.

BUG=543330

Review URL: https://codereview.chromium.org/1400073005

Cr-Original-Commit-Position: refs/heads/master@{#354175}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9c4323ea54d0d281b6c8a685d50b95af59e01d65