Devtools UI: Fix soft context menu alignment issues
This change fixes sub-menu alignment issues where the sub menu was
positioned before it was even sized, causing issues when it needed to be
on the left. Another issue that was fixed was caused by height scrollbars
causing the width to be insufficent.

This change also allows submenus to be moved up vertically if necessary,
removing the need for a scrollbar.

BUG=543963,544060,543454

Review URL: https://codereview.chromium.org/1410083002

Cr-Original-Commit-Position: refs/heads/master@{#354871}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a0266d94df56d543291baa3f1e2323bd63cafcb2