Invalidate background-attachment:fixed on scroll.
If there is any slow repaint objects (i.e. background-attachment:fixed),
we fall back to the FrameView::scrollContentsSlowPath() that invalidates
the whole main graphics layer. However this is not correct because the
object is not necessarily backed by the main backing. We should invalidate
them individually.

Note: This approach is still not entirely correct. With composited overflow
scrolling, we should inhibit threaded scrolling when a slow repaint descandant
presents and do similar invalidation. This is non-trivial to implement.
As the old compositing architecture is deprecated anyway, we will try to fix
the issue with SPv2.

BUG=540499

Review URL: https://codereview.chromium.org/1394193003

Cr-Original-Commit-Position: refs/heads/master@{#354960}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1d3aacc9cd5028db3a36036bfbace8933097be00