Check skip_script_comparison in ServiceWorkerWriteToCacheJob
When "Force update on page load" checkbox in DevTools is checked, we should skip
the script comparison and execute "install" and "activate" handler even if the
script is same as the existing SW. So we introduced skip_script_comparison flag.
https://crbug.com/500428#c19

But this change accidentally removed the flag check.
https://chromium.googlesource.com/chromium/src/+/4af0ccc451bb88e032e5c8003da2892bb1565a67

So this cl re-introduce the flag check.

And this cl changes the LayoutTest service-workers-force-update-on-page-load.html not to change the SW script.

BUG=500428

Review URL: https://codereview.chromium.org/1407123005

Cr-Original-Commit-Position: refs/heads/master@{#354986}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 68c242cab35167ed06650d80da101efbce44a6ae