Parse animation-name in CSSPropertyParser with CSSParserTokens
This patch moves the parsing logic for animation-name out of the
LegacyCSSProperty and into the CSSPropertyParser. This property has
special behaviour when specified as -webkit-animation-name, so
parseSingleValue is updated to take an unresolved properties instead
of a resolved property. consumeAnimationValue's arguments are extended
so we can continue counting usage of the non-standard behaviour.

BUG=499780

Review URL: https://codereview.chromium.org/1410223004

Cr-Original-Commit-Position: refs/heads/master@{#354993}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 271e821edda950d1f54af8bc3d6d8217ddfb0427