[css-grid] Fix freeSpace computation with non-zero baseSize flex tracks
The freeSpace computation was incorrect whenever the flex tracks had
a non-zero baseSize before the 1fr unit size resolution. That happened
because when assigning the new base size to the flex track, we were
unconditionally subtracting the whole base size to the freeSpace. That's
correct if the track is a "pure" flex track, i.e. 2fr, but if the track had
a non-zero base size (like minmax(10px, 1fr)) then both the new and the old
base sizes were incorrectly used to compute the freeSpace.

We can test the amount of remaining freeSpace by using content distribution to align and item place on a non-zero baseSize flex track. If the freeSpace computation is correct then the alignment will be accurate.

Review URL: https://codereview.chromium.org/1408403003

Cr-Original-Commit-Position: refs/heads/master@{#355028}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7eafefcbd32c3e1b0c534dc54694065926076f6b