Remove SelectorCheckingContext::m_scopeContainsLastMatchedElement.
With the change for SelectorChecker.cpp in
https://codereview.chromium.org/1245933002/
now we can assume that if SelectorCheckingContext has
non-null |scope|, the selector matching should take care
that the last matched element should be in the same scope
as |scope| always.

It means that m_scopeContainsLastMatchedElement in
SelectorCheckingContext and its derivative in
ElementRuleCollector are now redundant and can be
removed.

TEST=pass all existing tests
BUG=

Review URL: https://codereview.chromium.org/1360293003

Cr-Original-Commit-Position: refs/heads/master@{#355036}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9974f332f66d482645c3ad63d733cba55cc2178b