Fix LayoutBox::perpendicularContainingBlockLogicalHeight() when available logical height is not available
availableLogicalHeight() returns -1 when available logical height is
not available. LayoutBox::perpendicularContainingBlockLogicalHeight()
takes minimum of that without taking the possibility into account.

This patch fixes the function to return fillFallbackExtent (which is
the logical height of the frameView) in that case.

BUG=543647

Review URL: https://codereview.chromium.org/1419583003

Cr-Original-Commit-Position: refs/heads/master@{#355235}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fed69992a9de1357f400379e13fa87dbeaae14b4