Move WebImageGenerator to platform/exported.
It implements a public/platform/ header, not a public/web/ one. And it's
implemented atop Source/platform/, so it doesn't have any other need to be
in Source/web/.

This puts it in the same place as WebData and other similar cases.

Review URL: https://codereview.chromium.org/1421613002

Cr-Original-Commit-Position: refs/heads/master@{#355376}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9a021f104edefa759648517b9b7a6eaf3c3674e2