Never paginate fixed-positioned objects.
They are supposed to be repeated on every page anyway. Not that we currently do
that, but there's still no need to paginate them. It can only do harm.

Added a flexbox test, which is what the bug report was about. This regressed
for flexbox (and other types of non-block-container blocks as well) with
https://codereview.chromium.org/1360753002 , but this has actually been broken
for ages for regular block containers, so I threw in a bonus test for that too.

BUG=535163
R=leviw@chromium.org

Review URL: https://codereview.chromium.org/1426443003

Cr-Original-Commit-Position: refs/heads/master@{#356041}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 344e69ca2b1f2e20346eeb1f6d3d7d9fad38fd82