ContainerNode::getUpperLeftCorner should always call localToAbsolute.
This broke in http://crrev.com/1346043002 for the case of a LayoutText with no
InlineTextBox children (unrendered whitespace).

Arguably we should skip ALL empty LayoutText's when locating the LayoutObject to
base our position on (there is a check that skips them in some cases, poorly).
But this is the safer fix since it is closer to the M46 behavior.

I've filed http://crbug.com/547858 for the fix to skip all whitespace.

BUG=546478

Review URL: https://codereview.chromium.org/1429453002

Cr-Original-Commit-Position: refs/heads/master@{#356087}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bbe6dd0cfe21f22a294346aabaeea4d77d1fe299