Rename cdm_adapter to ppapi_cdm_adapter
As the current cdm_adapter is specific to ppapi, rename it to make it
clearer. This also moves some of the generic files out of the ppapi
folder so that they can be used by other implementations.

TEST=compiles

Review URL: https://codereview.chromium.org/1402413005

Cr-Original-Commit-Position: refs/heads/master@{#356196}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 38ae661b252d977ce675140c6e5470204809d0c0