Remove unnecessary check for hasVirtualLogicalHeight
We already check that condition in logicalHeight(). If, instead,
there's a performance reason to actually avoid the function call
to logicalHeight in the case of a virtualLogicalHeight, we should
split the check out of logicalHeight and have a logicalHeightInternal
method.

Review URL: https://codereview.chromium.org/1425713003

Cr-Original-Commit-Position: refs/heads/master@{#356213}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 797d517dc7dfe79df808094a5d914dcc6b776832