Replace HashMap<DatabaseGuid, HashSet<Database*>*> with HashCountedSet<DatabaseGuid>.
Before this CL, it had no mean to support uniqueness of
Database* in the HashSet, because elements were added in
only constructors.
And it was also useless to keep pointers to databases,
which were not used.

As a result, guidToDatabase() just counted the number of
databases used with each GUID.  So this CL replaces
HashMap<DatabaseGuid, HashSet<Database*>*> with HashCountedSet<DatabaseGuid>.

BUG=515524

Review URL: https://codereview.chromium.org/1424663002

Cr-Original-Commit-Position: refs/heads/master@{#356768}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5fb9cd6b34574457347b31ec163198f039673d70