Pass InterpolationEnvironment as a reference to InterpolationType conversion functions
This change converts the InterpolationEnvironment* parameter of
InterpolationType::maybeConvertPairwise() and InterpolationType::maybeConvertSingle()
into an InterpolationEnvironment&.
Previously we would attempt conversion of the keyframes without an environment
until crrev.com/70ab3999 which removed the optimisation.

This patch is to improve code health in the absence of this optimisation
with the reasoning that we can add the optimisation back given performance
data to justify the complexity cost.

Review URL: https://codereview.chromium.org/1428513003

Cr-Original-Commit-Position: refs/heads/master@{#356769}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 11fb42276c0ceff1f5e0ac9c3ff20b726692c8e0