Add UI for actions to Files app.
This CL adds support for actions (sharing, pinning, adding shortcuts, etc.) to
Files app, including custom actions provided via FSP API.

Note, that before we had sharing, pinning, and shortcuts hardcoded to Files
app for Drive only. Now, the code is refactored, and all of them became
so called actions. Some of them (sharing, pinning) are available to FSP
extensions. It means that FSP extensions can now support pinning and sharing,
as well as any other arbitrary actions.

This CL introduces a concept of actions with three following classes:
- ActionsModel, which represends a collection of actions for a set of entries.
- ActionsController, which reacts to user interaction, and creates
  ActionsModel instances when selection or focus is changed.
- ActionsSubmenu which is basically a view class responsible for rendering
  ActionsModel onto the context menu.

ActionsModel is not self updating, as FSP API doesn't support notifying
about changes of available actions. So, ActionsModel is recreated when there
is a chance that the list of actions might have changed. However, ActionsModel
dispatches an event 'invalidate' when the model requires recreation.

Since calculating actions is asynchronous (FSP API calls are asynchronous),
ActionsController computes ActionsModel for both - navigation list and
file list, so when a user requests a context menu, the actions are available
immediately, which avoids flicker of the context menu.

If a user right clicks on a file without selecting it earlier, then the menu
will flicker, as ActionsModel is not ready by the time the menu is shown.
It may cause the context menu partly invisible, as position of the menu is
calculated based on the menu items available at the time of showing the menu.
This will be addressed in a separate CL.

TEST=unit_tests: *ActionsSubmenu*, *ActionsModel*, many browser tests.
BUG=449525

Review URL: https://codereview.chromium.org/1383433002

Cr-Original-Commit-Position: refs/heads/master@{#356777}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a06edd06d28fa0a461da1d2e13948e863d26e3c1