stack: Adjust Pre-M Android incorrect debuggerd addresses.
Code location addresses printed by debuggerd in pre-M Android
are wrong where relocations are packed, because the libunwind
on these targets does not understand non-zero min vaddrs in
ELF LOAD segments.

Fix by searching the APK staging directory for libraries with
packed relocations, and adjusting the tombstone data on pre-
processing so that code addresses printed for these libraries
by debuggerd are offset by the min vaddr of the library that was
included in the APK.

This feature is hidden behind a new --packed-relocation-adjustments
flag, and defines a new --chrome-apk-dir flag that can be used
to indicate the location of the APK staging directory (defaults
to ../chrome_apk/ within the chrome-symbols-dir).

Uses the 'Build fingerprint' in the tombstone file to try to
auto-detect whether packed relocation adjustment should be
applied (that is, if the target is running pre-M Android).

BUG=536847

Review URL: https://codereview.chromium.org/1412223008

Cr-Original-Commit-Position: refs/heads/master@{#356858}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a0773c285a4a3a1a77df987a187d83096feb635f