Repaint on interest rect change
This is needed before we enable sync painting. Repaint when interest
rect changed sufficiently.

Also combine GraphicsLayer::paintIfNeeded() and GraphicsLayer::paint().
The former was needed because the caller needed to check the result to
determine if it's safe to commit the new paints. Now this is not needed.
Actually we always paint, and output CachedDisplayItemList when the
whole display item list is cached.

BUG=536999

Review URL: https://codereview.chromium.org/1428643004

Cr-Original-Commit-Position: refs/heads/master@{#356979}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f118da73a9f15420c1757133de7b89a019186bb3