Devtools Animations: Add method to fetch CSS keyframed animations
This change adds a method in StyleResolver which checks the animation-name
property of an element and returns all CSSOM keyframes rule representations
for that element. This method can then be used in InspectorCSSAgent to propogate
to the front-end and show the keyframes there.

BUG=266276

Review URL: https://codereview.chromium.org/1371723002

Cr-Original-Commit-Position: refs/heads/master@{#357041}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 981e28fe80483dc7edf3687205472c6bae61db79