Simplify TouchAction enum to be a simple bit flag
Originally I used the value 0 for the default touch-action value
(TouchActionAuto).  But as touch-action has expanded, it would be
a lot more convenient if we used a pure bit field (where "auto"
was all flags set, and "none" was 0).  This was hard to change
before the blink merge, but is now pretty easy.

BUG=None

Review URL: https://codereview.chromium.org/1422513006

Cr-Original-Commit-Position: refs/heads/master@{#357111}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a8b478daf9eef9e4f8e1ba21327c081b0301945e