Fix wrong usages of Vector<ContextLifecycleObserver*> and Vector<Observer*> on Oilpan builds.
To make those loops to iterate HeapHashSet<WeakMember<>> directly,
we have to confirm that no elements are removed in the iteration.
However, it is difficult to prove it with keeping flexibility of
other methods.
For example, we know some LifecycleObserver::contestDestroyed()
call LifecycleNotifier::removeObserver(), and it breaks the iteration
in LifecycleNotifier::notifyContextDestroyed().

BUG=515524

Review URL: https://codereview.chromium.org/1405303002

Cr-Original-Commit-Position: refs/heads/master@{#357328}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eea62bee00653bccd189ade6ece3571d106aedd5